Added linux as build target
This is the result of running `flutter create --platforms=linux .`
This command adds some changes to the pubspec.lock file but I was able
to revert those changes with no issue. It also add a unit test file
unter /test which I also removed.

Otherwise, this commit constists of just running the above command.